+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Not released
 
 - Fix executeSQL typing [#280](https://github.com/CartoDB/carto-react/pull/280)
+- Expose height and dense prop in TableWidget [#279](https://github.com/CartoDB/carto-react/pull/279)
+- Expose initialPageSize prop in TableWidget [#281](https://github.com/CartoDB/carto-react/pull/281)
 - Fix filtersToSQL output when IN filter has numeric values [#277](https://github.com/CartoDB/carto-react/pull/277)
 - Remove uniqueIdProperty default value in useGeojsonFeatures [#273](https://github.com/CartoDB/carto-react/pull/273)
 - Allow disable widgets filtering [#268](https://github.com/CartoDB/carto-react/pull/268)